requirements.txt
# Core dependencies
opencv-python-headless==4.11.0.86
numpy==1.26.4
pillow==10.4.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# API Framework
flask==3.0.0
flask-cors==4.0.0

# Database
# Using built-in sqlite3

# Video processing
requests==2.32.5

# Utilities
python-dateutil==2.8.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
